+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package grepros
 1.0.5 (2023-12-21)
 -------------------
 * fix ISO datetime support in earliest and latest timestamp arguments
+* fix api.dict_to_message() erroring on temporal types in dict
 
 1.0.4 (2023-09-28)
 -------------------